web: limit number of certificates created
LE integration is using a new account for each certificate created.
Therefore, it's easy to hit the account limit (10) when spawning a new
host. A manual workaround is to copy account_key.json from one vhost
to another. In this change, we just make a single certificate for
several vhosts.