isso: tentative to use systemd containers for Isso
This should have a better integration (sharing the store, not building
an image), but from a security point of view, I don't know exactly.
While the container itself is not "secure", Isso is running as a dynamic
user. This should be better than the Docker option (except Docker may
have more security around seccomp).